agregando anotaciones aclases y haciendo andar la APP
agregue @ID e hice  andar la app

+++ src/main/java/com/tiendapeliculas/tiendaPeliculas/data/actores/IActoresData.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 import com.tiendapeliculas.tiendaPeliculas.model.Actor;
 @Repository
-public interface IActoresData extends CrudRepository<Actor, Integer>{
+public interface IActoresData extends CrudRepository<Actor, Long>{
 
 }
